# recipes-back/src/main/resources/api/server/recipeApi-swagger.yaml
openapi: 3.0.0
info:
  title: Recipe Management
  description: >-
    This is Swagger UI environment generated for the TMF Recipe specification
  version: '4.0'
servers:
  - url: 'http://serverRoot/api/v1'
...
paths:
...
  '/recipes/{id}':
    get:
      tags:
        - recipe
      summary: Retrieves a recipe by ID
      description: This operation retrieves a recipe entity.
      operationId: retrieveRecipe
      parameters:
        - name: id
          in: path
          description: Identifier of the recipe
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/recipe'
        '400':
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - recipe
      summary: Updates partially a recipe
      description: This operation updates partially a recipe entity.
      operationId: updateRecipe
      parameters:
        - name: id
          in: path
          description: Identifier of the recipe
          required: true
          schema:
            type: string
      requestBody:
        description: The recipe to be updated
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/recipe'
        required: true
      responses:
        '200':
          description: Updated
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/recipe'
        '400':
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: Recipe
    delete:
      tags:
        - recipe
      summary: Delete a recipe
      description: This operation deletes a recipe entity.
      operationId: deleteRecipe
      parameters:
        - name: id
          in: path
          description: Identifier of the component ccomptcaa
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
          content: { }
        '400':
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
  /ingredients:
    get:
      tags:
        - ingredient
      summary: List or find ingredients
      description: This operation list or find ingredient entities
      operationId: listIngredients
      parameters:
        - name: fields
          in: query
          description: Comma-separated properties to be provided in response
          schema:
            type: string
        - name: offset
          in: query
          description: Requested index for start of resources to be provided in response
          schema:
            type: integer
        - name: limit
          in: query
          description: Requested number of resources to be provided in response
          schema:
            type: integer
        - name: name
          in: query
          description: Name of the ingredient to be provided in response
          schema:
            type: string
        - name: recipeId
          in: query
          description: Id of the recipe to be provided in response
          schema:
            type: integer
      responses:
        '200':
          description: Success
          headers:
            X-Total-Count:
              description: Total number of items matching criteria
              schema:
                type: integer
            X-Result-Count:
              description: Actual number of items returned in the response body
              schema:
                type: integer
          content:
            application/json;charset=utf-8:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ingredient'
        '400':
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - ingredient
      summary: creates an ingredient
      description: This operation creates an ingredient entity.
      operationId: createIngredient
      requestBody:
        description: The ingredient
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/ingredient'
        required: true
      responses:
        '200':
          description: Updated
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ingredient'
        '400':
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: Ingredient
  '/ingredients/{id}':
    get:
      tags:
        - ingredient
      summary: Retrieves an ingredient by ID
      description: This operation retrieves a ingredient entity.
      operationId: retrieveIngredient
      parameters:
        - name: id
          in: path
          description: Identifier of the ingredient
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ingredient'
        '400':
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - ingredient
      summary: Updates partially a ingredient
      description: This operation updates partially a ingredient entity.
      operationId: updateIngredient
      parameters:
        - name: id
          in: path
          description: Identifier of the ingredient
          required: true
          schema:
            type: string
      requestBody:
        description: The ingredient to be updated
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/ingredient'
        required: true
      responses:
        '200':
          description: Updated
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ingredient'
        '400':
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: Ingredient
    delete:
      tags:
        - ingredient
      summary: Delete a ingredient
      description: This operation deletes a ingredient entity.
      operationId: deleteIngredient
      parameters:
        - name: id
          in: path
          description: Identifier of the ingredient
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
          content: { }
        '400':
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    recipe:
      type: object
      properties:
        id:
          type: integer
          description: id of the recipe
        name:
          type: string
          description: name of the recipe
        category:
          type: string
          description: category of the recipe
          format: string
        image/png:
          type: string
          format: binary
          description: image of the ingredient
        url:
          type: string
          description: url of the recipe
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/ingredient'
    ingredient:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              description: id of the ingredient
            name:
              type: string
              description: name of the ingredient
            measure:
              type: string
              description: measure of the ingredient
            image/png:
              type: string
              format: binary
              description: image of the ingredient
            subtitle:
              type: string
              description: subtitles, if present of the ingredient
            recipe_id:
              type: integer
...
    Error:
      required:
        - code
        - reason
      type: object
      properties:
        code:
          type: string
          description: 'Application relevant detail, defined in the API or a common list.'
        reason:
          type: string
          description: >-
            Explanation of the reason for the error which can be shown to a
            client user.
        message:
          type: string
          description: >-
            More details and corrective actions related to the error which can
            be shown to a client user.
        status:
          type: string
          description: HTTP Error code extension
        referenceError:
          type: string
          description: URI of documentation describing the error.
          format: uri
        atBaseType:
          type: string
          description: 'When sub-classing, this defines the super-class.'
        atSchemaLocation:
          type: string
          description: >-
            A URI to a JSON-Schema file that defines additional attributes and
            relationships
          format: uri
        atType:
          type: string
          description: 'When sub-classing, this defines the sub-class entity name.'
        detail:
          type: object
          properties: { }
          description: 'When sub-classing, this defines the sub-class entity name.'
      description: >-
        Used when an API throws an Error, typically with a HTTP error
        response-code (3xx, 4xx, 5xx)